// packages/fhir-r5/src/v5.0.0/measurestratifier/types.ts
import type {
  Extension,
  Element,
  CodeableConcept,
  Expression,
  Reference,
} from "../core/types";
import type { MeasureComponent } from "../measurecomponent/types";

/** Generated from FHIR JSON Schema */

/** The Measure resource provides the definition of a quality measure. */
export interface MeasureStratifier {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  linkId?: string;

  _linkId?: Element;

  code?: CodeableConcept;

  description?: string;

  _description?: Element;

  criteria?: Expression;

  groupDefinition?: Reference;

  component?: MeasureComponent[];
}